resolve sim-uav circular dependencies
