+++ controls/sae_2025_ws/src/uav/uav/utils.py
@@ -86,47 +86,6 @@ def find_folder_with_heuristic(folder_name, home_dir=None, keywords=("penn", "ai
     return find_folder(folder_name, home_dir)
 
 
-def px4_path_from_gitmodules(gitmodules):
-    # Return the declared submodule path for PX4-Autopilot from .gitmodules in case PX4-Autopilot location shifts.
-    sections = []
-    current = None
-    for raw in Path(gitmodules).read_text().splitlines():
-        line = raw.strip()
-        if line.startswith("[submodule"):
-            current = {}
-            sections.append(current)
-        elif current is not None and "=" in line:
-            key, _, value = line.partition("=")
-            current[key.strip().lower()] = value.strip()
-    for section in sections:
-        path = section.get("path", "")
-        url = section.get("url", "")
-        if "px4-autopilot" in path.lower() or "px4-autopilot" in url.lower():
-            return path or None
-    return None
-
-
-def get_px4_submodule_path():
-    # Resolve the in-repo PX4-Autopilot submodule path via the repo's .gitmodules
-    # Returns None when the submodule doesn't exist or isn't initialized
-    here = Path(__file__).resolve()
-    for parent in here.parents:
-        gitmodules = parent / ".gitmodules"
-        if gitmodules.is_file():
-            rel_path = px4_path_from_gitmodules(gitmodules)
-            if rel_path:
-                candidate = (parent / rel_path).resolve()
-                # Non-empty confirms the submodule is checked out, not a placeholder.
-                if candidate.is_dir() and any(candidate.iterdir()):
-                    return str(candidate)
-            return None
-    return None
-
-
-# Prefer submodule path if available, otherwise fall back to a user-local px4
-DEFAULT_PX4_PATH = get_px4_submodule_path() or os.path.expanduser("~/PX4-Autopilot")
-
-
 def get_airframe_details(px4_path, airframe_id):
     """
     Parses PX4 airframe files to find vehicle type and model name from an ID.
